requirements.txt
future
pillow
websocket_client
six